+++ src/components/date/date.jsx
@@ -10,11 +10,13 @@ import React, {
 import calendarSvg from '../../svgs/calendar.svg';
 import cautionSvg from '../../svgs/caution.svg';
 import Calendar from '../calendar/calendar.jsx';
+import Control from '../control/control.jsx';
 import FormControl from '../form-control/form-control.jsx';
 import IconButton from '../icon-button/icon-button.jsx';
 import Icon from '../icon/icon.jsx';
 import useDropdownClose from '../../hooks/use-dropdown-close.js';
 import styles from './date.css';
+import useForwardedRef from '../../hooks/use-forwarded-ref.js';
 
 /* eslint-disable react/prop-types */
 function ControlIcons(props) {
@@ -64,10 +66,7 @@ function fromFullDateFormat(string) {
 }
 
 const Date = forwardRef(function Date(props, forwardedRef) {
-  const backupRef = useRef();
-  const ref = forwardedRef || backupRef;
-  const containerRef = useRef();
-  const inputRef = useRef();
+  const ref = useForwardedRef(forwardedRef);
   const [active, setActive] = useState(false);
   const [editing, setEditing] = useState(!!props.validationMessage);
   const [expanded, setExpanded] = useState(false);
@@ -88,20 +87,25 @@ const Date = forwardRef(function Date(props, forwardedRef) {
     tooltip: `${props.id}-tooltip`,
     validationMessage: `${props.id}Hint`,
   };
+  const refs = {
+    container: useRef(),
+    control: useRef(),
+    input: useRef(),
+  };
 
   function onBlur(event) {
-    if (containerRef.current.contains(event.relatedTarget)) return;
+    if (refs.container.current.contains(event.relatedTarget)) return;
 
     // Given the validation spec, any provided validation messages are cleared on blur.
     // If the native node is still invalid then re-render with the native validation messages.
-    inputRef.current.setCustomValidity('');
-    setValidationMessage(inputRef.current.validationMessage);
+    refs.input.current.setCustomValidity('');
+    setValidationMessage(refs.input.current.validationMessage);
     setActive(false);
-    if (!inputRef.current.validationMessage) setEditing(false);
+    if (!refs.input.current.validationMessage) setEditing(false);
   }
 
   function onInputChange() {
-    setValue(fromFullDateFormat(inputRef.current.value));
+    setValue(fromFullDateFormat(refs.input.current.value));
   }
 
   function onInputClick(event) {
@@ -139,7 +143,7 @@ const Date = forwardRef(function Date(props, forwardedRef) {
     setExpanded(false);
   }
 
-  useDropdownClose(containerRef, expanded, () => setExpanded(false));
+  useDropdownClose(refs.container, expanded, () => setExpanded(false));
 
   useEffect(() => {
     setEditing(!!props.validationMessage);
@@ -155,24 +159,23 @@ const Date = forwardRef(function Date(props, forwardedRef) {
   }, [value]);
 
   useLayoutEffect(() => {
-    if (active) inputRef.current.focus();
+    if (active) refs.input.current.focus();
   }, [editing]);
 
   useLayoutEffect(() => {
-    inputRef.current.setCustomValidity(validationMessage);
+    refs.input.current.setCustomValidity(validationMessage);
   }, [validationMessage]);
 
   useLayoutEffect(() => {
-    inputRef.current.value = editing ? toFullDateFormat(value) : toDateStringFormat(value);
+    refs.input.current.value = editing ? toFullDateFormat(value) : toDateStringFormat(value);
   }, [value]);
 
   useImperativeHandle(ref, () => ({
+    ...refs.control.current,
     get dirty() {
       const providedValue = props.value;
       return providedValue !== this.value;
     },
-    id: props.id,
-    name: props.name,
     get value() {
       return toFullDateFormat(value);
     },
@@ -182,41 +185,51 @@ const Date = forwardRef(function Date(props, forwardedRef) {
     <div
       className={classNames.layout.container}
       onBlur={onBlur}
-      ref={containerRef}
+      ref={refs.container}
     >
       <FormControl
         error={validationMessage}
         id={ids.input}
-        labelId={ids.label}
         label={props.label}
+        labelId={ids.label}
+        name={props.name}
         readOnly={props.readOnly}
+        ref={refs.control}
         required={props.required}
         tooltip={props.tooltip}
       >
-        <input
-          aria-describedby={`${ids.validationMessage} ${ids.tooltip}`}
-          className={classNames.input}
-          defaultValue={editing ? toFullDateFormat(value) : toDateStringFormat(value)}
-          id={ids.input}
-          key={`${ids.input}-${editing ? 'editing' : 'display'}`}
-          max={props.max}
-          min={props.min}
-          name={props.name}
-          onChange={onInputChange}
-          onClick={onInputClick}
-          onKeyDown={onInputKeyDown}
-          placeholder={props.placeholder}
-          readOnly={props.readOnly}
-          ref={inputRef}
-          required={props.required}
-          type={editing ? 'date' : 'text'}
-        />
-        <ControlIcons
-          label={props.label}
-          onPress={onIconPress}
-          readOnly={props.readOnly}
+        <Control
+          labelledBy={ids.label}
           validationMessage={validationMessage}
-        />
+          validationMessageId={ids.validationMessage}
+        >
+          <div style={{ position: 'relative' }}>
+            <input
+              aria-describedby={`${ids.validationMessage} ${ids.tooltip}`}
+              className={classNames.input}
+              defaultValue={editing ? toFullDateFormat(value) : toDateStringFormat(value)}
+              id={ids.input}
+              key={`${ids.input}-${editing ? 'editing' : 'display'}`}
+              max={props.max}
+              min={props.min}
+              name={props.name}
+              onChange={onInputChange}
+              onClick={onInputClick}
+              onKeyDown={onInputKeyDown}
+              placeholder={props.placeholder}
+              readOnly={props.readOnly}
+              ref={refs.input}
+              required={props.required}
+              type={editing ? 'date' : 'text'}
+            />
+            <ControlIcons
+              label={props.label}
+              onPress={onIconPress}
+              readOnly={props.readOnly}
+              validationMessage={validationMessage}
+            />
+          </div>
+        </Control>
       </FormControl>
       {expanded && (
         <div className={classNames.layout.calendar}>