+++ src/components/form-control/form-control.test.jsx
@@ -5,32 +5,19 @@ import FormControl from './form-control.jsx';
 
 describe('<FormControl>', () => {
   const requiredProps = {
-    children: <input id="test-id" />,
+    children: <input aria-describedby="test-idHint" id="test-id" />,
     id: 'test-id',
     label: 'Test label',
   };
 
   it('has defaults', () => {
-    render(<FormControl {...requiredProps} />);
+    render(<FormControl {...requiredProps}>{requiredProps.children}</FormControl>);
     expect(document.body).toMatchSnapshot();
-  });
-
-  it('requires either an id or a labelId', () => {
-    /* eslint-disable no-console */
-    const originalConsoleError = console.error;
-    console.error = jest.fn();
-
-    render(<FormControl {...requiredProps} id={undefined} labelId={undefined} />);
-    expect(console.error).toHaveBeenNthCalledWith(1, 'Warning: Failed prop type: Invalid prop `id` supplied to `FormControl`. Either `id` or `labelId` are required.\n    in FormControl');
-    expect(console.error).toHaveBeenNthCalledWith(2, 'Warning: Failed prop type: Invalid prop `labelId` supplied to `FormControl`. Either `id` or `labelId` are required.\n    in FormControl');
-
-    console.error = originalConsoleError;
-    /* eslint-enable no-console */
   });
 
   describe('className API', () => {
     it('can be set', () => {
-      render(<FormControl {...requiredProps} className="unique-class" />);
+      render(<FormControl {...requiredProps} className="unique-class">{requiredProps.children}</FormControl>);
       expect(screen.getByLabelText('Test label').parentElement.parentElement).toHaveAttribute('class', 'unique-class');
     });
   });
@@ -40,7 +27,7 @@ describe('<FormControl>', () => {
       const inputRef = createRef();
       render((
         <FormControl {...requiredProps}>
-          <input id="test-id" ref={inputRef} />
+          <input aria-describedby="test-idHint" id="test-id" ref={inputRef} />
         </FormControl>
       ));
       expect(screen.getByLabelText('Test label')).toBe(inputRef.current);
@@ -49,18 +36,8 @@ describe('<FormControl>', () => {
 
   describe('error API', () => {
     it('can be set', () => {
-      render(<FormControl {...requiredProps} error="I am an error message" />);
-      expect(screen.getByText('I am an error message')).toBeInTheDocument();
-    });
-
-    it('uses aria-live polite to alert users of an error', () => {
-      render(<FormControl {...requiredProps} error="I am an error message" />);
-      expect(screen.getByText('I am an error message')).toHaveAttribute('aria-live', 'polite');
-    });
-
-    it('uses a hint to alert users of the error message', () => {
-      render(<FormControl {...requiredProps} error="I am an error message" />);
-      expect(screen.getByText('I am an error message')).toHaveAttribute('id', 'test-idHint');
+      render(<FormControl {...requiredProps} error="I am an error message">{requiredProps.children}</FormControl>);
+      expect(document.body).toMatchSnapshot();
     });
   });
 
@@ -68,7 +45,7 @@ describe('<FormControl>', () => {
     it('can be set without a labelId', () => {
       render((
         <FormControl {...requiredProps} id="unique-id">
-          <input id="unique-id" />
+          <input aria-describedby="unique-idHint" id="unique-id" />
         </FormControl>
       ));
       expect(screen.getByLabelText('Test label')).toHaveAttribute('id', 'unique-id');
@@ -77,14 +54,14 @@ describe('<FormControl>', () => {
 
   describe('label API', () => {
     it('can be set', () => {
-      render(<FormControl {...requiredProps} label="Unique label" />);
+      render(<FormControl {...requiredProps} label="Unique label">{requiredProps.children}</FormControl>);
       expect(screen.getByLabelText('Unique label')).toBeInTheDocument();
     });
   });
 
   describe('labelId API', () => {
-    it('can be set without an id', () => {
-      render(<FormControl {...requiredProps} id={undefined} labelId="unique-label-id" />);
+    it('can be set', () => {
+      render(<FormControl {...requiredProps} labelId="unique-label-id">{requiredProps.children}</FormControl>);
       expect(screen.getByText('Test label')).toHaveAttribute('id', 'unique-label-id');
     });
   });
@@ -92,33 +69,21 @@ describe('<FormControl>', () => {
   describe('onKeyDown API', () => {
     it('can be set', () => {
       const onKeyDownSpy = jest.fn();
-      render(<FormControl {...requiredProps} onKeyDown={onKeyDownSpy} />);
+      render(<FormControl {...requiredProps} onKeyDown={onKeyDownSpy}>{requiredProps.children}</FormControl>);
       user.click(screen.getByText('Test label'));
       user.keyboard('a');
       expect(onKeyDownSpy).toHaveBeenCalled();
     });
   });
 
-  describe('readOnly API', () => {
-    it('can be set and hides the error state', () => {
-      render(<FormControl {...requiredProps} readOnly={true} error="I am invalid" />);
-      expect(screen.queryByText('I am invalid')).not.toBeInTheDocument();
-    });
-
-    it('can be unset', () => {
-      render(<FormControl {...requiredProps} readOnly={false} error="I am invalid" />);
-      expect(screen.getByText('I am invalid')).toBeInTheDocument();
-    });
-  });
-
   describe('required API', () => {
     it('can be set', () => {
-      render(<FormControl {...requiredProps} required={true} />);
+      render(<FormControl {...requiredProps} required={true}>{requiredProps.children}</FormControl>);
       expect(screen.getByText('(Required)')).toBeInTheDocument();
     });
 
     it('can be unset', () => {
-      render(<FormControl {...requiredProps} required={false} />);
+      render(<FormControl {...requiredProps} required={false}>{requiredProps.children}</FormControl>);
       expect(screen.queryByText('(Required)')).not.toBeInTheDocument();
     });
   });