+++ src/components/multi-select/multi-select.jsx
@@ -7,6 +7,7 @@ import React, {
   forwardRef,
 } from 'react';
 import PropTypes from 'prop-types';
+import Control from '../control/control.jsx';
 import FormControl from '../form-control/form-control.jsx';
 import FormControlIcons from '../form-control-icons/form-control-icons.jsx';
 import Icon from '../icon/icon.jsx';
@@ -21,6 +22,7 @@ import TagList from '../tag-list/tag-list.jsx';
 import Tag from '../tag/tag.jsx';
 import useDropdownClose from '../../hooks/use-dropdown-close.js';
 import styles from './multi-select.css';
+import useForwardedRef from '../../hooks/use-forwarded-ref.js';
 
 function getFormControlChildrenContainerClassName(readOnly, validationMessage) {
   if (readOnly) {
@@ -36,16 +38,13 @@ function getFormControlChildrenContainerClassName(readOnly, validationMessage) {
 
 const MultiSelect = forwardRef(function MultiSelect(props, ref) {
   const [autocompleteValue, setAutocompleteValue] = useState('');
-  const autocompleteRef = useRef();
-  const backupRef = useRef();
   const [expanded, setExpanded] = useState(false);
-  const selfRef = ref || backupRef;
+  const selfRef = useForwardedRef(ref);
   const [validationMessage, setValidationMessage] = useState(props.validationMessage);
   const [value, setValue] = useState(props.value || []);
   const valueRefs = value.map(() => createRef());
   const visibleOptions = getVisibleOptions();
   const visibleOptionsRefs = visibleOptions.map(() => createRef());
-  const wrapperRef = useRef(null);
 
   const ids = {
     emptyMessage: `${props.id}-empty`,
@@ -53,6 +52,12 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
     listbox: `${props.id}-listbox`,
     textbox: `${props.id}-autocomplete`,
     tooltip: `${props.id}-autocomplete-tooltip`,
+    validation: `${props.id}Hint`,
+  };
+  const refs = {
+    autocomplete: useRef(),
+    control: useRef(),
+    wrapper: useRef(),
   };
 
   const classNames = {
@@ -126,11 +131,11 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
   }
 
   function onAutocompleteBlur() {
-    if (!autocompleteRef.current) return;
+    if (!refs.autocomplete.current) return;
 
-    autocompleteRef.current.setCustomValidity('');
-    if (!autocompleteRef.current.validity.valid) {
-      setValidationMessage(autocompleteRef.current.validationMessage);
+    refs.autocomplete.current.setCustomValidity('');
+    if (!refs.autocomplete.current.validity.valid) {
+      setValidationMessage(refs.autocomplete.current.validationMessage);
     } else {
       setValidationMessage(props.validationMessage || '');
     }
@@ -159,7 +164,7 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
     if (!props.readOnly) {
       setValue([]);
       setExpanded(false);
-      autocompleteRef.current.focus();
+      refs.autocomplete.current.focus();
     }
   }
 
@@ -171,7 +176,7 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
       .sort((a, b) => props.optionIDGetter(a) - props.optionIDGetter(b));
     setValue(newValue);
     setAutocompleteValue('');
-    autocompleteRef.current.focus();
+    refs.autocomplete.current.focus();
   }
 
   function onClick(event) {
@@ -190,7 +195,7 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
     }
   }
 
-  useDropdownClose(wrapperRef, expanded, onDropdownClose);
+  useDropdownClose(refs.wrapper, expanded, onDropdownClose);
 
   useEffect(() => {
     setValidationMessage(props.validationMessage);
@@ -205,89 +210,97 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
   }, [value.length]);
 
   useImperativeHandle(selfRef, () => ({
+    ...refs.control.current,
     get dirty() {
       return props.value.map(val => JSON.stringify(val)).join(',') !== this.value.map(val => JSON.stringify(val)).join(',');
     },
     id: props.id,
-    name: props.name,
     get value() {
       return value;
     },
   }));
 
   return (
-    <div className={classNames.container} id={props.id} ref={wrapperRef}>
+    <div className={classNames.container} id={props.id} ref={refs.wrapper}>
       <FormControl
         error={validationMessage}
         id={ids.textbox}
         label={props.label}
         labelId={ids.label}
+        name={props.name}
         onKeyDown={onKeyDown}
+        ref={refs.control}
         required={props.required}
         tooltip={props.tooltip}
       >
-        <div role="presentation" className={classNames.formControlChildrenContainer} onClick={onClick}>
-          <TagList
-            className={classNames.tagList}
-            labelledBy={ids.label}
-            refs={valueRefs}
-          >
-            {value.length !== 0 &&
-            props.tagChildren ?
-              props.tagChildren(value, valueRefs, onOptionRemove) :
-              value.map((val, index) => (
-                <Tag
-                  defaultActive={index === 0}
-                  id={`${props.id}-option-${props.optionIDGetter(val)}`}
-                  key={props.optionIDGetter(val)}
-                  onRemove={onOptionRemove}
-                  readOnly={props.readOnly}
-                  ref={valueRefs[index]}
-                >
-                  {props.optionLabelGetter(val)}
-                </Tag>
-              ))
-            }
-            <input
-              aria-autocomplete="list"
-              aria-controls={ids.listbox}
-              aria-describedby={`${ids.emptyMessage} ${ids.tooltip}`}
-              aria-expanded={expanded}
-              aria-haspopup="listbox"
-              autoComplete="off"
-              role="combobox"
-              className={classNames.input}
-              id={ids.textbox}
-              onBlur={onAutocompleteBlur}
-              onChange={onAutocompleteChange}
-              onInput={props.onInput}
-              placeholder={value.length === 0 ? props.placeholder : undefined}
-              readOnly={props.readOnly}
-              required={props.required ? value.length === 0 : false}
-              ref={autocompleteRef}
-              value={autocompleteValue}
-            />
-          </TagList>
-          <FormControlIcons validationMessage={validationMessage} className={classNames.iconsContainer}>
-            {(!props.readOnly && value.length > 0) && (
-              <IconButton
-                icon={iconClear}
-                label={`Remove all selected options on ${props.label}`}
-                onPress={onOptionsClear}
-                className={classNames.iconClear}
+        <Control
+          labelledBy={ids.label}
+          validationMessage={validationMessage || ''}
+          validationMessageId={ids.validation}
+        >
+          <div role="presentation" className={classNames.formControlChildrenContainer} onClick={onClick}>
+            <TagList
+              className={classNames.tagList}
+              labelledBy={ids.label}
+              refs={valueRefs}
+            >
+              {value.length !== 0 &&
+              props.tagChildren ?
+                props.tagChildren(value, valueRefs, onOptionRemove) :
+                value.map((val, index) => (
+                  <Tag
+                    defaultActive={index === 0}
+                    id={`${props.id}-option-${props.optionIDGetter(val)}`}
+                    key={props.optionIDGetter(val)}
+                    onRemove={onOptionRemove}
+                    readOnly={props.readOnly}
+                    ref={valueRefs[index]}
+                  >
+                    {props.optionLabelGetter(val)}
+                  </Tag>
+                ))
+              }
+              <input
+                aria-autocomplete="list"
+                aria-controls={ids.listbox}
+                aria-describedby={`${ids.emptyMessage} ${ids.tooltip}`}
+                aria-expanded={expanded}
+                aria-haspopup="listbox"
+                autoComplete="off"
+                role="combobox"
+                className={classNames.input}
+                id={ids.textbox}
+                onBlur={onAutocompleteBlur}
+                onChange={onAutocompleteChange}
+                onInput={props.onInput}
+                placeholder={value.length === 0 ? props.placeholder : undefined}
+                readOnly={props.readOnly}
+                required={props.required ? value.length === 0 : false}
+                ref={refs.autocomplete}
+                value={autocompleteValue}
               />
-            )}
-            {props.readOnly ? (
-              <Icon icon={iconCaretDown} label={`Opening ${props.label} options disabled while read only`} />
-            ) : (
-              <IconButton
-                icon={iconCaretDown}
-                label={`Open ${props.label} options`}
-                onPress={onClick}
-              />
-            )}
-          </FormControlIcons>
-        </div>
+            </TagList>
+            <FormControlIcons validationMessage={validationMessage} className={classNames.iconsContainer}>
+              {(!props.readOnly && value.length > 0) && (
+                <IconButton
+                  icon={iconClear}
+                  label={`Remove all selected options on ${props.label}`}
+                  onPress={onOptionsClear}
+                  className={classNames.iconClear}
+                />
+              )}
+              {props.readOnly ? (
+                <Icon icon={iconCaretDown} label={`Opening ${props.label} options disabled while read only`} />
+              ) : (
+                <IconButton
+                  icon={iconCaretDown}
+                  label={`Open ${props.label} options`}
+                  onPress={onClick}
+                />
+              )}
+            </FormControlIcons>
+          </div>
+        </Control>
       </FormControl>
       {dropdownContents()}
     </div>