+++ src/components/percentage/percentage.jsx
@@ -5,6 +5,7 @@ import React, {
   useLayoutEffect,
   useRef,
 } from 'react';
+import Control from '../control/control.jsx';
 import FormControl from '../form-control/form-control.jsx';
 import FormControlIcons from '../form-control-icons/form-control-icons.jsx';
 import styles from './percentage.css';
@@ -18,16 +19,22 @@ function getClassName(className, validationMessage) {
 const Percentage = forwardRef(function Percentage(props, forwardedRef) {
   const fallbackRef = useRef();
   const ref = forwardedRef || fallbackRef;
-  const inputRef = useRef();
-  const [validationMessage, validate] = useValidation(props.validationMessage, inputRef);
+
   const classNames = {
     container: props.cssContainer,
     input: props.className,
   };
   const ids = {
+    label: `${props.id}-label`,
     validationMessage: `${props.id}Hint`,
     tooltip: `${props.id}-tooltip`,
   };
+  const refs = {
+    control: useRef(),
+    input: useRef(),
+  };
+
+  const [validationMessage, validate] = useValidation(props.validationMessage, refs.input);
 
   function onBlur() {
     validate();
@@ -38,18 +45,17 @@ const Percentage = forwardRef(function Percentage(props, forwardedRef) {
   }
 
   useLayoutEffect(() => {
-    inputRef.current.value = props.value || '';
+    refs.input.current.value = props.value || '';
   }, [props.value]);
 
   useImperativeHandle(ref, () => ({
+    ...refs.control.current,
     get dirty() {
       const providedValue = props.value || undefined;
       return !Object.is(providedValue, this.value);
     },
-    id: props.id,
-    name: props.name,
     get value() {
-      return window.parseFloat(inputRef.current.value) || undefined;
+      return window.parseFloat(refs.input.current.value) || undefined;
     },
   }));
 
@@ -59,30 +65,41 @@ const Percentage = forwardRef(function Percentage(props, forwardedRef) {
       error={validationMessage}
       id={props.id}
       label={props.label}
+      labelId={ids.label}
+      name={props.name}
       readOnly={props.readOnly}
+      ref={refs.control}
       required={props.required}
       tooltip={props.tooltip}
     >
-      <input
-        aria-describedby={`${ids.validationMessage} ${ids.tooltip}`}
-        className={getClassName(classNames.input, validationMessage)}
-        defaultValue={props.value}
-        id={props.id}
-        max={100}
-        min={0}
-        name={props.name}
-        onBlur={onBlur}
-        onChange={onChange}
-        placeholder={props.placeholder}
-        readOnly={props.readOnly}
-        ref={inputRef}
-        required={props.required}
-        step={0.01}
-        type="number"
-      />
-      <FormControlIcons validationMessage={validationMessage} className={styles['icons-container']}>
-        <span className={styles['percent-sign']}>%</span>
-      </FormControlIcons>
+      <Control
+        labelledBy={ids.label}
+        validationMessage={validationMessage}
+        validationMessageId={ids.validationMessage}
+      >
+        <div style={{ position: 'relative' }}>
+          <input
+            aria-describedby={`${ids.validationMessage} ${ids.tooltip}`}
+            className={getClassName(classNames.input, validationMessage)}
+            defaultValue={props.value}
+            id={props.id}
+            max={100}
+            min={0}
+            name={props.name}
+            onBlur={onBlur}
+            onChange={onChange}
+            placeholder={props.placeholder}
+            readOnly={props.readOnly}
+            ref={refs.input}
+            required={props.required}
+            step={0.01}
+            type="number"
+          />
+          <FormControlIcons validationMessage={validationMessage} className={styles['icons-container']}>
+            <span className={styles['percent-sign']}>%</span>
+          </FormControlIcons>
+        </div>
+      </Control>
     </FormControl>
   );
 });